+++ Engine/Renderer/RenderContext.inl
@@ -4,7 +4,7 @@
 template<typename Command>
 Command *RenderContext::create()
 {
-	void *packet = packets.alloc(CommandPacket::getSize<Command>());
+	void *packet = packets.alloc((uint32_t)CommandPacket::getSize<Command>());
 	CommandPacket::setSubmitCallback(packet, Command::submit);
 
 	return CommandPacket::getCommand<Command>(packet);
